% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_ammendment_number}
\alias{add_ammendment_number}
\title{Adds the amendment number of an rcv to the rcv meta data}
\usage{
add_ammendment_number(x)
}
\description{
Adds the amendment number of an rcv to the rcv meta data
}
