% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{character_date_format}
\alias{character_date_format}
\title{Catches user input errors for character stop dates}
\usage{
character_date_format(usr_date, date_format = "\%d-\%m-\%Y")
}
\description{
\code{character_date_format()} checks for date format entry in the form
"\%d-\%m-\%y".
}
