% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean column names}
\usage{
clean_column_names(data_set)
}
\description{
Function that cleans column names by lower casing them, removing trailing
full stops, replaces repeated full stops with one full stop and finally
replaces full stops with underscores.
}
