% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_title_doc_old.r
\name{extract_title_doc_old}
\alias{extract_title_doc_old}
\title{Extracts titles from EP 7 and older documents}
\usage{
extract_title_doc_old(page)
}
\description{
Extracts titles from EP 7 and older documents
}
