% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_colspan}
\alias{get_colspan}
\title{Finds the colspan of each cell
xml_nodeset The table nodeset. An xml2 nodeset.}
\usage{
get_colspan(xml_nodeset, max_columns, ns)
}
\description{
Finds the colspan of each cell
xml_nodeset The table nodeset. An xml2 nodeset.
}
