% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{level1_info_fun}
\alias{level1_info_fun}
\title{function extracts all information organised in a table tag with
the class: "doc_subtitle_level1"}
\usage{
level1_info_fun(page)
}
\description{
function extracts all information organised in a table tag with
the class: "doc_subtitle_level1"
}
