% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{level1_lead_extract}
\alias{level1_lead_extract}
\title{function extracts lead committee and lead rapporteur from level1_info
and removes these entries from level1_info}
\usage{
level1_lead_extract(x)
}
\description{
function extracts lead committee and lead rapporteur from level1_info
and removes these entries from level1_info
}
