% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{level2_assoc_extract}
\alias{level2_assoc_extract}
\title{function extracts associated committee(s) and associated rapporteur(s)}
\usage{
level2_assoc_extract(x)
}
\description{
function extracts associated committee(s) and associated rapporteur(s)
}
