% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{level2_lead_extract}
\alias{level2_lead_extract}
\title{function extracts lead committee and lead rapporteur from level2_info
and removes these entries from level2_info}
\usage{
level2_lead_extract(x)
}
\description{
function extracts lead committee and lead rapporteur from level2_info
and removes these entries from level2_info
}
