% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_title}
\alias{match_title}
\title{Matches the position of titles within the xml tree structure
relative to all occurences of "//w:p|//w:tbl"}
\usage{
match_title(string, pattern)
}
\description{
Matches the position of titles within the xml tree structure
relative to all occurences of "//w:p|//w:tbl"
}
