% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vote_results.r
\name{parse_vote_results}
\alias{parse_vote_results}
\title{Call to parsing functions for vote results tables
The function "parse_vote_results" assigns a class to current vote and calls
paring function.}
\usage{
parse_vote_results(x, data_dir, verbose)
}
\arguments{
\item{x}{current row of m_data}

\item{data_dir}{inherits from parse_votes}

\item{verbose}{inherits from parse_votes}
}
\value{
placeholder.
}
\description{
Call to parsing functions for vote results tables
The function "parse_vote_results" assigns a class to current vote and calls
paring function.
}
\examples{
\dontrun{
# Not yet
}

}
