% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_votes.R
\name{parse_votes}
\alias{parse_votes}
\title{Parse vote data from a directory}
\usage{
parse_votes(
  data_dir,
  overwrite = FALSE,
  verbose = TRUE,
  start_date = NULL,
  end_date = NULL,
  cores = 1
)
}
\arguments{
\item{data_dir}{The path to the directory containing the vote data.
A character string.}

\item{overwrite}{A logical value indicating whether to overwrite existing
parsed data.}

\item{verbose}{A logical value indicating whether to print progress messages.}

\item{start_date}{The start date for filtering the vote data. A character
string.}

\item{end_date}{The end date for filtering the vote data. A character string.}

\item{cores}{The number of CPU cores to use for parallel processing.
An integer.}
}
\value{
Writes parsed vote data to the parsed/votes.RData file, votes_meta
data to the parsed/votes_meta.RData file, and vote requests data to the
parsed/vote_requests.RData file.
}
\description{
This function parses vote data from a directory and saves the parsed data.
}
\examples{
parse_votes(
 data_dir = "C:/Users/Philipp/Documents/EP 7/",
 overwrite = FALSE,
 verbose = TRUE,
 start_date = "14-07-2009",
 end_date = "30-06-2014",
 cores = 13)
}
