% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scrape_combined_rcvs}
\alias{scrape_combined_rcvs}
\title{Scrapes RC documents.}
\usage{
scrape_combined_rcvs(data_dir, verbose = TRUE, sleep = 1, links)
}
\description{
RC documents are used to reference roll-call votes that combine several
requests into one roll-call.
}
