% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_descriptions}
\alias{search_descriptions}
\title{Searches for the descriptions of the votes below the title}
\usage{
search_descriptions(page, t_pos)
}
\description{
Searches for the descriptions of the votes below the title
}
