% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_votes.r
\name{search_votes.xml}
\alias{search_votes.xml}
\title{This is the code for xml votes up until 2024. The EP may change this
in the future. For new xml code, use the function search_votes.xml_v2}
\usage{
\method{search_votes}{xml}(x, data_dir, verbose, ...)
}
\description{
This is the code for xml votes up until 2024. The EP may change this
in the future. For new xml code, use the function search_votes.xml_v2
}
