% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_votes.r
\name{search_votes.xml_v2}
\alias{search_votes.xml_v2}
\title{As of March 2024, this function works for the xml files from 2024.
All other xmls are parsed using search_votes.xml}
\usage{
\method{search_votes}{xml_v2}(x, data_dir, verbose, ...)
}
\description{
As of March 2024, this function works for the xml files from 2024.
All other xmls are parsed using search_votes.xml
}
